Created hook use-draw-on-image for image drawing logic
